+++ src/b-map/resources/js/Pages/Reviews/Edit.vue
@@ -34,6 +34,7 @@ const submitReview = () => {
     <Layout :error="error">
         <div class="flex items-center">
             <img src="/images/marker_icon.png" class="h-12 w-8 mx-1 pt-4">
+
             <h1 class="text-xl font-bold pt-4">{{ review.spot.name }}</h1>   
         </div>
           
@@ -42,6 +43,8 @@ const submitReview = () => {
         <h2 class="text-lg font-bold ml-7 my-6">編集する</h2>
         
         <p class="ml-7">評価</p>
+
+        <!-- デフォルトのフォーム送信動作を無効化してカスタムメソッドを実行させる -->
         <form @submit.prevent="submitReview">
             <div class="ml-6 mt-2">
                 <span v-for="n in 5" :key="n" class="star cursor-pointer text-gray-500 text-4xl ml-1" @click="rating = n" :class="{ 'active': n <= rating }">&#9733;</span>
@@ -51,8 +54,11 @@ const submitReview = () => {
             
             <div class="flex flex-col">
                 <textarea v-model="comment" placeholder="レビューコメント（任意）" class="mt-8 min-h-32"></textarea>
+
                 <p v-if="errors.comment" class="text-red-500 mt-2 m-auto">※{{ errors.comment }}</p>
+
                 <BlueButton type="submit" class="w-40 m-auto mt-14">レビューを更新</BlueButton>
+                
                 <p v-if="errors.spot_id" class="text-red-500 mt-5 m-auto">※{{ errors.spot_id }}</p>
             </div>
         </form>
